requirements.txt
astropy==6.0.1
einops==0.8.1
matplotlib==3.8.0
neurokit2==0.2.7
numpy==2.3.2
opencv_contrib_python==4.10.0.84
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
Pillow==11.3.0
pytz==2023.4
scikit_learn==1.7.1
scipy==1.16.0
torch==2.2.2
torch==2.5.1
torchvision==0.20.1
torchvision==0.17.2
tqdm==4.67.1
tqdm==4.65.2
